requirements.txt
Fabric3
jinja2
six
